backend/pyproject.toml
[tool.poetry]
name = "end-to-end-encrypt-chat"
version = "0.1.0"
description = ""
authors = ["mirzaev.arsen <mirzaev.a@cadesign.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = {extras = ["all"], version = "^0.133.0"}
pydantic-settings = "^2.13.1"
uvicorn = {extras = ["standard"], version = "^0.41.0"}
gunicorn = "^25.1.0"
sqlalchemy = "^2.0.47"
psycopg2-binary = "^2.9.11"
alembic = "^1.18.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.5.0"
python-dotenv = "^1.2.1"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"